requirements.txt
alembic==1.3.2
cffi==1.13.2
Click==7.0
cryptography==2.8
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
nltk==3.4.5
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==3.11.2
python-socketio==4.4.0
six==1.14.0
SQLAlchemy==1.3.12
visitor==0.1.3
Werkzeug==0.16.0
